@@ -31,7 +31,7 @@ class _EditUserState extends State<EditUser> {
   TextEditingController password = new TextEditingController();
   TextEditingController description = new TextEditingController();
 
-  TextEditingController hospital = new TextEditingController();
+  //TextEditingController hospital = new TextEditingController();
 
   TextEditingController city = new TextEditingController();
   String region = "north";
@@ -43,10 +43,11 @@ class _EditUserState extends State<EditUser> {
   }
 
   Future<void> _pickImage() async {
-    File selectedFile = await FilePicker.getFile(type: FileType.image);
+    FilePickerResult selectedFile =
+        await FilePicker.platform.pickFiles(type: FileType.image);
     if (selectedFile != null && this.mounted) {
       setState(() {
-        profilePic = selectedFile;
+        profilePic = File(selectedFile.files.first.path);
       });
     }
   }
@@ -89,24 +90,21 @@ class _EditUserState extends State<EditUser> {
 
     if (profilePic != null) {
       try {
-        StorageReference storageReference = FirebaseStorage.instance
-            .ref()
-            .child('AvanaFiles/profilepics/' + profilePic.path.split("/").last);
-        StorageUploadTask uploadTask = storageReference.putFile(profilePic);
-        await uploadTask.onComplete;
-        profilePickUrl = await storageReference.getDownloadURL();
+        profilePickUrl = await Utils.uploadImageGetUrl(
+            'AvanaFiles/profilepics/' + profilePic.path.split("/").last,
+            profilePic);
       } catch (Exception) {}
     }
 
-    await Firestore.instance
+    await FirebaseFirestore.instance
         .collection("userdata")
-        .document(currentUserId)
-        .updateData({
+        .doc(currentUserId)
+        .update({
       "password": password.text,
       "isactive": isActiveUser,
       "userrole": userRole,
       "description": description.text,
-      "hospital": hospital.text,
+      //"hospital": hospital.text,
       "city": city.text,
       "region": region,
       "profile_pic_url": profilePickUrl,
@@ -121,15 +119,15 @@ class _EditUserState extends State<EditUser> {
   }
 
   Future<void> fetchUserDetails() async {
-    DocumentSnapshot currentUserDetails = await Firestore.instance
+    DocumentSnapshot currentUserDetails = await FirebaseFirestore.instance
         .collection('userdata')
-        .document(currentUserId)
+        .doc(currentUserId)
         .get();
 
     currentUserEmail = currentUserDetails["email"];
     isActiveUser = currentUserDetails["isactive"];
     password.text = currentUserDetails["password"];
-    hospital.text = currentUserDetails["hospital"];
+    //hospital.text = currentUserDetails["hospital"];
     city.text = currentUserDetails["city"];
     region = currentUserDetails["region"];
     description.text = currentUserDetails["description"];
@@ -162,9 +160,9 @@ class _EditUserState extends State<EditUser> {
                                   actions: [
                                     FlatButton(
                                       onPressed: () => {
-                                        Firestore.instance
+                                        FirebaseFirestore.instance
                                             .collection('userdata')
-                                            .document(currentUserId)
+                                            .doc(currentUserId)
                                             .delete(),
                                         Navigator.pop(context),
                                         Navigator.pop(context),
@@ -218,14 +216,14 @@ class _EditUserState extends State<EditUser> {
                             contentPadding:
                                 EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
                             labelText: "Password")),
-                    SizedBox(height: 10),
+                    /*SizedBox(height: 10),
                     TextField(
                         controller: hospital,
                         decoration: InputDecoration(
                             border: OutlineInputBorder(),
                             contentPadding:
                                 EdgeInsets.fromLTRB(20.0, 15.0, 20.0, 15.0),
-                            labelText: "Hospital Name")),
+                            labelText: "Hospital Name")),*/
                     SizedBox(height: 10),
                     TextField(
                         controller: city,
@@ -309,7 +307,7 @@ class _EditUserState extends State<EditUser> {
                       ),
                     ),
                     ListTile(
-                      title: new Text('Faculty'),
+                      title: new Text(Utils.distributorName),
                       leading: Radio(
                         value: 2,
                         groupValue: userRole,
