@@ -26,6 +26,16 @@ class _FeedPageState extends State<FeedPage> {
     setState(() {});
   }
 
+  List<Widget> getFeeds(QuerySnapshot snapshot) {
+    List<Widget> result = List<Widget>.empty();
+
+    snapshot.docs.forEach((document) {
+      //result.add();
+    });
+
+    return result;
+  }
+
   Widget build(BuildContext context) {
     medQry = MediaQuery.of(context);
     return WillPopScope(
@@ -42,7 +52,7 @@ class _FeedPageState extends State<FeedPage> {
             title: Text('Feed'),
           ),
           body: new StreamBuilder<QuerySnapshot>(
-            stream: Firestore.instance
+            stream: FirebaseFirestore.instance
                 .collection('feed')
                 .orderBy("created_time", descending: true)
                 .limit(300)
@@ -53,133 +63,138 @@ class _FeedPageState extends State<FeedPage> {
                 return SizedBox(
                     child: new LinearProgressIndicator(), height: 5);
               return new ListView(
-                children: snapshot.data.documents.map((document) {
-                  return GestureDetector(
-                      onTap: () => {
-                            Navigator.pushNamed(context, "/feeddetails",
-                                arguments: document.documentID)
-                          },
-                      child: new Card(
-                        child: new Padding(
-                            padding: EdgeInsets.all(10),
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              mainAxisAlignment: MainAxisAlignment.start,
-                              children: [
-                                document["feedtype"] == 1
-                                    ? new Container(
-                                        child: RichText(
-                                          text: TextSpan(
-                                            text: document["ownername"],
-                                            style: TextStyle(
-                                                fontWeight: FontWeight.bold,
-                                                color: Colors.black,
-                                                fontSize: 16),
-                                            children: <TextSpan>[
-                                              TextSpan(
-                                                  text: document["content"],
-                                                  style: TextStyle(
-                                                      fontWeight:
-                                                          FontWeight.normal,
-                                                      color: Colors.black)),
-                                            ],
-                                          ),
-                                        ),
-                                      )
-                                    : new Container(
-                                        child: Column(
-                                          crossAxisAlignment:
-                                              CrossAxisAlignment.start,
-                                          mainAxisAlignment:
-                                              MainAxisAlignment.start,
-                                          children: [
-                                            Padding(
-                                                padding: EdgeInsets.fromLTRB(
-                                                    5, 0, 0, 0),
-                                                child: Row(children: [
-                                                  Utils.userProfilePic(
-                                                      document["owner"], 12),
-                                                  SizedBox(
-                                                    width: 7,
-                                                  ),
-                                                  Text(document["ownername"],
-                                                      style: TextStyle(
-                                                          fontWeight:
-                                                              FontWeight.bold,
-                                                          color: Colors.black,
-                                                          fontSize: 18))
-                                                ])),
-                                            SizedBox(
-                                              height: 5,
-                                            ),
-                                            Padding(
-                                                padding: EdgeInsets.fromLTRB(
-                                                    5, 0, 0, 0),
-                                                child: Text(
-                                                    Utils.getMessageTimerFrmt(
-                                                        document[
-                                                            "created_time"]),
-                                                    style: TextStyle(
-                                                        color: Colors.black45,
-                                                        fontSize: 12))),
-                                            SizedBox(
-                                              height: 10,
-                                            ),
-                                            Padding(
-                                                padding: EdgeInsets.fromLTRB(
-                                                    5, 0, 0, 0),
-                                                child: Text(document["content"],
+                children: snapshot.data.docs.map((document) {
+                  if (document.id == null || document["ownername"] == null) {
+                    return SizedBox();
+                  } else {
+                    return GestureDetector(
+                        onTap: () => {
+                              Navigator.pushNamed(context, "/feeddetails",
+                                  arguments: document.id)
+                            },
+                        child: new Card(
+                          child: new Padding(
+                              padding: EdgeInsets.all(10),
+                              child: Column(
+                                crossAxisAlignment: CrossAxisAlignment.start,
+                                mainAxisAlignment: MainAxisAlignment.start,
+                                children: [
+                                  document["feedtype"] == 1
+                                      ? new Container(
+                                          child: RichText(
+                                            text: TextSpan(
+                                              text: document["ownername"],
+                                              style: TextStyle(
+                                                  fontWeight: FontWeight.bold,
+                                                  color: Colors.black,
+                                                  fontSize: 16),
+                                              children: <TextSpan>[
+                                                TextSpan(
+                                                    text: document["content"],
                                                     style: TextStyle(
                                                         fontWeight:
                                                             FontWeight.normal,
-                                                        fontSize: 15,
-                                                        color: Colors.black))),
-                                            document["attachments"].length > 0
-                                                ? new Container(
-                                                    padding:
-                                                        EdgeInsets.fromLTRB(
-                                                            0, 5, 0, 5),
-                                                    height: 100,
-                                                    width:
-                                                        medQry.size.width * .88,
-                                                    child: Utils
-                                                        .attachmentPreviewSlider(
-                                                            context,
-                                                            document,
-                                                            null))
-                                                : SizedBox()
-                                          ],
+                                                        color: Colors.black)),
+                                              ],
+                                            ),
+                                          ),
+                                        )
+                                      : new Container(
+                                          child: Column(
+                                            crossAxisAlignment:
+                                                CrossAxisAlignment.start,
+                                            mainAxisAlignment:
+                                                MainAxisAlignment.start,
+                                            children: [
+                                              Padding(
+                                                  padding: EdgeInsets.fromLTRB(
+                                                      5, 0, 0, 0),
+                                                  child: Row(children: [
+                                                    Utils.userProfilePic(
+                                                        document["owner"], 12),
+                                                    SizedBox(
+                                                      width: 7,
+                                                    ),
+                                                    Text(document["ownername"],
+                                                        style: TextStyle(
+                                                            fontWeight:
+                                                                FontWeight.bold,
+                                                            color: Colors.black,
+                                                            fontSize: 18))
+                                                  ])),
+                                              SizedBox(
+                                                height: 5,
+                                              ),
+                                              Padding(
+                                                  padding: EdgeInsets.fromLTRB(
+                                                      5, 0, 0, 0),
+                                                  child: Text(
+                                                      Utils.getMessageTimerFrmt(
+                                                          document[
+                                                              "created_time"]),
+                                                      style: TextStyle(
+                                                          color: Colors.black45,
+                                                          fontSize: 12))),
+                                              SizedBox(
+                                                height: 10,
+                                              ),
+                                              Padding(
+                                                  padding: EdgeInsets.fromLTRB(
+                                                      5, 0, 0, 0),
+                                                  child: Text(
+                                                      document["content"],
+                                                      style: TextStyle(
+                                                          fontWeight:
+                                                              FontWeight.normal,
+                                                          fontSize: 15,
+                                                          color:
+                                                              Colors.black))),
+                                              document["attachments"].length > 0
+                                                  ? new Container(
+                                                      padding:
+                                                          EdgeInsets.fromLTRB(
+                                                              0, 5, 0, 5),
+                                                      height: 100,
+                                                      width: medQry.size.width *
+                                                          .88,
+                                                      child: Utils
+                                                          .attachmentPreviewSlider(
+                                                              context,
+                                                              document,
+                                                              null))
+                                                  : SizedBox()
+                                            ],
+                                          ),
                                         ),
-                                      ),
-                                SizedBox(
-                                  height: 10,
-                                ),
-                                Row(children: [
-                                  document["attachments"].length > 0
-                                      ? Text(
-                                          document["attachments"]
-                                                  .length
-                                                  .toString() +
-                                              " Attachment",
-                                          style: TextStyle(fontSize: 14),
-                                        )
-                                      : SizedBox(),
-                                  Spacer(),
-                                  Utils.feedCommentCount
-                                          .containsKey(document.documentID)
-                                      ? Text(
-                                          Utils.feedCommentCount[
-                                                      document.documentID]
-                                                  .toString() +
-                                              " "
-                                                  "Comment",
-                                          style: TextStyle(fontSize: 14),
-                                        )
-                                      : SizedBox(),
-                                ]),
-                              ],
-                            )),
-                      ));
+                                  SizedBox(
+                                    height: 10,
+                                  ),
+                                  Row(children: [
+                                    document["attachments"].length > 0
+                                        ? Text(
+                                            document["attachments"]
+                                                    .length
+                                                    .toString() +
+                                                " Attachment",
+                                            style: TextStyle(fontSize: 14),
+                                          )
+                                        : SizedBox(),
+                                    Spacer(),
+                                    Utils.feedCommentCount
+                                            .containsKey(document.id)
+                                        ? Text(
+                                            Utils.feedCommentCount[document.id]
+                                                    .toString() +
+                                                " "
+                                                    "Comment",
+                                            style: TextStyle(fontSize: 14),
+                                          )
+                                        : SizedBox(),
+                                  ]),
+                                ],
+                              )),
+                        ));
+                  }
                 }).toList(),
               );
             },
@@ -194,7 +209,7 @@ class _FeedPageState extends State<FeedPage> {
             onTap: _onItemTapped,
           ),
           floatingActionButton: new Visibility(
-              visible: true,
+              visible: Utils.userRole == 1,
               child: FloatingActionButton(
                 onPressed: () {
                   Navigator.pushNamed(context, "/feededitor");
